+++ src/core/constants/input.ts
@@ -16,3 +16,37 @@ export const defaultGamepadMapping = {
   14: 'left',
   15: 'right',
 }
+
+/**
+
+input_player1_a = "x"
+input_player1_b = "z"
+input_player1_down = "down"
+input_player1_gun_trigger_mbtn = "1"
+input_player1_joypad_index = "0"
+input_player1_l = "q"
+input_player1_left = "left"
+input_player1_mouse_index = "0"
+input_player1_r = "w"
+input_player1_right = "right"
+input_player1_select = "rshift"
+input_player1_start = "enter"
+input_player1_up = "up"
+input_player1_x = "s"
+input_player1_y = "a"
+
+ */
+export const defaultKeyboardMapping = {
+  z: 'b',
+  x: 'a',
+  a: 'y',
+  s: 'x',
+  q: 'l1',
+  w: 'r1',
+  rshift: 'select',
+  enter: 'start',
+  up: 'up',
+  down: 'down',
+  left: 'left',
+  right: 'right',
+}