+++ src/core/exposed/get-default-keyboard-mappings.ts
@@ -1,0 +1,5 @@
+import { defaultKeyboardMapping } from '../constants/input'
+
+export function getDefaultKeyboardMappings() {
+  return defaultKeyboardMapping
+}