+++ src/core/exposed/get-keyboard-mappings.ts
@@ -1,0 +1,5 @@
+import { PreferenceParser } from '../classes/preference-parser'
+
+export function getKeyboardMappings() {
+  return PreferenceParser.get('keyboardMappings')
+}